+++ apps/api/src/db/utils.ts
